requirements.txt
discord==2.3.2
openai==0.28.0
pymongo==4.5.0
python-dotenv==1.0.0
discord.py==2.4.0